Align number columns in table to the right
Aligning columns that contain numbers to the right slightly improves
readability, at least when number with one digit and two digits are
mixed.

In version 0.7.2 PrettyTable ignores the flag when you pass a default
alignment when constructing the table, so the alignment is set
explicitly for each column for now. This bug has already been fixed in
the trunk version of PrettyTable, but there is no new release yet.